requirements.txt
openai>=1.30.0
psycopg[binary]>=3.1
pgvector>=0.2.5
python-dotenv>=1.0.1

